moving from_* methods and related utilities in circuit templates sub modules
