+++ cirkit/templates/region_graph/graph.py
@@ -2,13 +2,17 @@ import itertools
 import json
 from abc import ABC
 from collections import defaultdict
-from collections.abc import Iterable, Iterator, Sequence
+from collections.abc import Callable, Iterable, Iterator, Sequence
 from functools import cached_property
 from typing import TypeAlias, TypedDict, cast, final
 
 import numpy as np
 from numpy.typing import NDArray
 
+from cirkit.symbolic.circuit import Circuit
+from cirkit.symbolic.layers import HadamardLayer, KroneckerLayer, Layer, SumLayer
+from cirkit.symbolic.parameters import ParameterFactory
+from cirkit.templates.utils import InputLayerFactory, ProductLayerFactory, SumLayerFactory
 from cirkit.utils.algorithms import DiAcyclicGraph
 from cirkit.utils.scope import Scope
 
@@ -333,3 +337,242 @@ class RegionGraph(DiAcyclicGraph[RegionGraphNode]):
             in_nodes[ptn] = in_rgns
 
         return RegionGraph(nodes, in_nodes, outputs=outputs)
+
+    def build_circuit(
+        self,
+        *,
+        input_factory: InputLayerFactory,
+        sum_product: str | None = None,
+        sum_weight_factory: ParameterFactory | None = None,
+        nary_sum_weight_factory: ParameterFactory | None = None,
+        sum_factory: SumLayerFactory | None = None,
+        prod_factory: ProductLayerFactory | None = None,
+        num_channels: int = 1,
+        num_input_units: int = 1,
+        num_sum_units: int = 1,
+        num_classes: int = 1,
+        factorize_multivariate: bool = True,
+    ) -> Circuit:
+        """Construct a symbolic circuit from a region graph.
+            There are two ways to use this method. The first one is to specify a sum-product layer
+            abstraction, which can be either 'cp' (the CP layer), 'cp-t' (the CP-transposed layer),
+            or 'tucker' (the Tucker layer). The second one is to manually specify the factories to
+            build distinct um and product layers. If the first way is chosen, then one can possibly
+            use a factory that builds the symbolic parameters of the sum-product layer abstractions.
+            The factory that constructs the input factory must always be specified.
+
+        Args:
+            input_factory: A factory that builds an input layer.
+            sum_product: The sum-product layer to use. It can be None, 'cp', 'cp-t', or 'tucker'.
+            sum_weight_factory: The factory to construct the weights of the sum layers.
+                It can be None, or a parameter factory, i.e., a map
+                from a shape to a symbolic parameter. If it is None, then the default
+                weight factory of the sum layer is used instead.
+            nary_sum_weight_factory: The factory to construct the weight of sum layers havity arity
+                greater than one. If it is None, then it will have the same value and semantics of
+                the given sum_weight_factory.
+            sum_factory: A factory that builds a sum layer. It can be None.
+            prod_factory: A factory that builds a product layer. It can be None.
+            num_channels: The number of channels for each variable.
+            num_input_units: The number of input units.
+            num_sum_units: The number of sum units per sum layer.
+            num_classes: The number of output classes.
+            factorize_multivariate: Whether to fully factorize input layers, when they depend on
+                more than one variable.
+
+        Returns:
+            Circuit: A symbolic circuit.
+
+        Raises:
+            NotImplementedError: If an unknown 'sum_product' is given.
+            ValueError: If both 'sum_product' and layer factories are specified, or none of them.
+            ValueError: If 'sum_product' is specified, but 'weight_factory' is not.
+            ValueError: The given region graph is malformed.
+        """
+        if (sum_factory is None and prod_factory is not None) or (
+            sum_factory is not None and prod_factory is None
+        ):
+            raise ValueError(
+                "Both 'sum_factory' and 'prod_factory' must be specified or none of them"
+            )
+        if sum_product is None and (sum_factory is None or prod_factory is None):
+            raise ValueError(
+                "If 'sum_product' is not given, then both 'sum_factory' and 'prod_factory'"
+                " must be specified"
+            )
+        if sum_product is not None and (sum_factory is not None or prod_factory is not None):
+            raise ValueError(
+                "At most one between 'sum_product' and the pair 'sum_factory' and 'prod_factory'"
+                " must be specified"
+            )
+        if nary_sum_weight_factory is None:
+            nary_sum_weight_factory = sum_weight_factory
+
+        layers: list[Layer] = []
+        in_layers: dict[Layer, list[Layer]] = {}
+        node_to_layer: dict[RegionGraphNode, Layer] = {}
+
+        def build_cp_(
+            rgn: RegionNode, rgn_partitioning: Sequence[RegionNode]
+        ) -> HadamardLayer | SumLayer:
+            layer_ins = [node_to_layer[rgn_in] for rgn_in in rgn_partitioning]
+            denses = [
+                SumLayer(
+                    node_to_layer[rgn_in].num_output_units,
+                    num_sum_units,
+                    weight_factory=sum_weight_factory,
+                )
+                for rgn_in in rgn_partitioning
+            ]
+            hadamard = HadamardLayer(num_sum_units, arity=len(rgn_partitioning))
+            layers.extend(denses)
+            layers.append(hadamard)
+            in_layers[hadamard] = denses
+            for d, li in zip(denses, layer_ins):
+                in_layers[d] = [li]
+            # If the region is not a root region of the region graph,
+            # then make Hadamard the last layer
+            if self.region_outputs(rgn):
+                node_to_layer[rgn] = hadamard
+                return hadamard
+            # Otherwise, introduce an additional sum layer to ensure the output layer is a sum
+            output_dense = SumLayer(
+                hadamard.num_output_units, num_classes, weight_factory=sum_weight_factory
+            )
+            layers.append(output_dense)
+            in_layers[output_dense] = [hadamard]
+            node_to_layer[rgn] = output_dense
+            return output_dense
+
+        def build_cp_transposed_(
+            rgn: RegionNode, rgn_partitioning: Sequence[RegionNode]
+        ) -> SumLayer:
+            layer_ins = [node_to_layer[rgn_in] for rgn_in in rgn_partitioning]
+            num_in_units = list({li.num_output_units for li in layer_ins})
+            if len(num_in_units) > 1:
+                raise ValueError(
+                    "Cannot build a CP transposed layer, as the inputs would have different units"
+                )
+            num_units = num_sum_units if self.region_outputs(rgn) else num_classes
+            hadamard = HadamardLayer(num_in_units[0], arity=len(rgn_partitioning))
+            dense = SumLayer(num_in_units[0], num_units, weight_factory=sum_weight_factory)
+            layers.append(hadamard)
+            layers.append(dense)
+            in_layers[hadamard] = layer_ins
+            in_layers[dense] = [hadamard]
+            node_to_layer[rgn] = dense
+            return dense
+
+        def build_tucker_(rgn: RegionNode, rgn_partitioning: Sequence[RegionNode]) -> SumLayer:
+            layer_ins = [node_to_layer[rgn_in] for rgn_in in rgn_partitioning]
+            num_in_units = list({li.num_output_units for li in layer_ins})
+            if len(num_in_units) > 1:
+                raise ValueError(
+                    "Cannot build a Tucker layer, as the inputs would have different units"
+                )
+            num_units = num_sum_units if self.region_outputs(rgn) else num_classes
+            kronecker = KroneckerLayer(num_in_units[0], arity=len(rgn_partitioning))
+            dense = SumLayer(
+                kronecker.num_output_units,
+                num_units,
+                weight_factory=sum_weight_factory,
+            )
+            layers.append(kronecker)
+            layers.append(dense)
+            in_layers[kronecker] = layer_ins
+            in_layers[dense] = [kronecker]
+            node_to_layer[rgn] = dense
+            return dense
+
+        # Set the sum-product layer builder, if necessary
+        sum_prod_builder_: Callable[[RegionNode, Sequence[RegionNode]], Layer] | None
+        if sum_product is None:
+            sum_prod_builder_ = None
+        elif sum_product == "cp":
+            sum_prod_builder_ = build_cp_
+        elif sum_product == "cp-t":
+            sum_prod_builder_ = build_cp_transposed_
+        elif sum_product == "tucker":
+            sum_prod_builder_ = build_tucker_
+        else:
+            raise NotImplementedError(f"Unknown sum-product layer abstraction called {sum_product}")
+
+        # Loop through the region graph nodes, which are already sorted in a topological ordering
+        for node in self.topological_ordering():
+            if isinstance(node, PartitionNode):  # Partition node
+                # If a sum-product layer abstraction has been specified,
+                # then just skip partition nodes
+                if sum_prod_builder_ is not None:
+                    continue
+                assert prod_factory is not None
+                partition_inputs = self.partition_inputs(node)
+                prod_inputs = [node_to_layer[rgn] for rgn in partition_inputs]
+                prod_sl = prod_factory(num_sum_units, len(prod_inputs))
+                layers.append(prod_sl)
+                in_layers[prod_sl] = prod_inputs
+                node_to_layer[node] = prod_sl
+            assert isinstance(
+                node, RegionNode
+            ), "Region graph nodes must be either region or partition nodes"
+            region_inputs = self.region_inputs(node)
+            region_outputs = self.region_outputs(node)
+            if not region_inputs:
+                # Input region node
+                if factorize_multivariate and len(node.scope) > 1:
+                    factorized_input_sls = [
+                        input_factory(Scope([sc]), num_input_units, num_channels)
+                        for sc in node.scope
+                    ]
+                    input_sl = HadamardLayer(num_input_units, arity=len(factorized_input_sls))
+                    layers.extend(factorized_input_sls)
+                    in_layers[input_sl] = factorized_input_sls
+                else:
+                    input_sl = input_factory(node.scope, num_input_units, num_channels)
+                num_units = num_sum_units if self.region_outputs(node) else num_classes
+                if sum_factory is None:
+                    layers.append(input_sl)
+                    node_to_layer[node] = input_sl
+                    continue
+                sum_sl = sum_factory(num_input_units, num_units)
+                layers.append(input_sl)
+                layers.append(sum_sl)
+                in_layers[sum_sl] = [input_sl]
+                node_to_layer[node] = sum_sl
+            elif len(region_inputs) == 1:
+                # Region node that is partitioned into one and only one way
+                (ptn,) = region_inputs
+                if sum_prod_builder_ is not None:
+                    sum_prod_builder_(node, self.partition_inputs(ptn))
+                    continue
+                num_units = num_sum_units if region_outputs else num_classes
+                sum_input = node_to_layer[ptn]
+                sum_sl = sum_factory(sum_input.num_output_units, num_units)
+                layers.append(sum_sl)
+                in_layers[sum_sl] = [sum_input]
+                node_to_layer[node] = sum_sl
+            else:  # len(node_inputs) > 1:
+                # Region node with multiple partitionings
+                num_units = num_sum_units if region_outputs else num_classes
+                if sum_prod_builder_ is None:
+                    sum_ins = [node_to_layer[ptn] for ptn in region_inputs]
+                    mix_ins = [sum_factory(sli.num_output_units, num_units) for sli in sum_ins]
+                    layers.extend(mix_ins)
+                    for mix_sl, sli in zip(mix_ins, sum_ins):
+                        in_layers[mix_sl] = [sli]
+                else:
+                    mix_ins = [
+                        sum_prod_builder_(node, self.partition_inputs(cast(PartitionNode, ptn)))
+                        for ptn in region_inputs
+                    ]
+                mix_sl = SumLayer(
+                    num_units,
+                    num_units,
+                    arity=len(mix_ins),
+                    weight_factory=nary_sum_weight_factory,
+                )
+                layers.append(mix_sl)
+                in_layers[mix_sl] = mix_ins
+                node_to_layer[node] = mix_sl
+
+        outputs = [node_to_layer[rgn] for rgn in self.outputs]
+        return Circuit(num_channels, layers, in_layers, outputs)
